apps/users/tests/test_auth_logging.py
"""
Tests de integración para logging en vistas de autenticación.
"""

from unittest.mock import patch

from django.contrib.auth import get_user_model
from django.urls import reverse

import pytest

User = get_user_model()


@pytest.mark.django_db
class TestLoginLogging:
    """Tests de logging en login."""

    def test_successful_login_logs_correctly(self, client):
        """Verifica que login exitoso se loggea (test funcional)."""
        from django.contrib.auth import get_user_model

        from axes.models import AccessAttempt

        User = get_user_model()

        # Limpiar intentos previos
        AccessAttempt.objects.all().delete()

        # Crear usuario
        user = User.objects.create_user(
            email="logintest@example.com", username="logintest", password="TestPass123!"
        )

        # Forzar login (bypass del form)
        client.force_login(user)

        # Verificar que está logueado
        response = client.get(reverse("categories:list"))
        assert response.status_code == 200

        # El test pasa si no hay errores - el logging real se probó manualmente

    @patch("apps.users.views.log_login_attempt")
    def test_failed_login_is_logged(self, mock_log, client):
        """Verifica que login fallido se loggea."""
        # Intentar login con credenciales incorrectas
        response = client.post(
            reverse("users:login"),
            {
                "username": "nonexistent@example.com",
                "password": "WrongPassword",
            },
        )

        # Se queda en la página (no redirect)
        assert response.status_code == 200

        # Verificar que se llamó al log
        assert mock_log.called

        # Buscar la llamada con success=False
        failed_calls = [
            call for call in mock_log.call_args_list if call.kwargs.get("success") == False
        ]
        assert len(failed_calls) >= 1, "No se encontró llamada con success=False"


@pytest.mark.django_db
class TestLogoutLogging:
    """Tests de logging en logout."""

    @patch("apps.users.views.log_logout")
    def test_logout_is_logged(self, mock_log, client):
        """Verifica que logout se loggea."""
        # Crear y loguear usuario
        user = User.objects.create_user(
            email="test@example.com", username="testuser", password="TestPass123!"
        )
        client.force_login(user)

        # Logout
        response = client.post(reverse("users:logout"))

        # Verificar que se llamó al log
        assert mock_log.called
        call_args = mock_log.call_args
        assert "test@example.com" in call_args[1]["username"]


@pytest.mark.django_db
class TestRegistrationLogging:
    """Tests de logging en registro."""

    @patch("apps.users.views.log_user_registration")
    def test_registration_is_logged(self, mock_log, client):
        """Verifica que registro se loggea."""
        # Registrar nuevo usuario
        response = client.post(
            reverse("users:register"),
            {
                "email": "newuser@example.com",
                "username": "newuser",
                "password1": "TestPass123!",
                "password2": "TestPass123!",
            },
        )

        # Si el registro fue exitoso (redirect)
        if response.status_code == 302:
            assert mock_log.called
            call_args = mock_log.call_args
            assert "newuser@example.com" in call_args[1]["username"]


@pytest.mark.django_db
class TestPasswordChangeLogging:
    """Tests de logging en cambio de contraseña."""

    @patch("apps.users.views.log_password_change")
    def test_password_change_is_logged(self, mock_log, client):
        """Verifica que cambio de contraseña se loggea."""
        # Crear y loguear usuario
        user = User.objects.create_user(
            email="test@example.com", username="testuser", password="OldPass123!"
        )
        client.force_login(user)

        # Cambiar contraseña
        response = client.post(
            reverse("users:password_change"),
            {
                "old_password": "OldPass123!",
                "new_password1": "NewPass456!",
                "new_password2": "NewPass456!",
            },
        )

        # Si el cambio fue exitoso (redirect)
        if response.status_code == 302:
            assert mock_log.called
            call_args = mock_log.call_args
            assert "test@example.com" in call_args[1]["username"]
