requirements/dev.txt
-r base.txt
django-debug-toolbar>=4.2
black>=24.0
flake8>=7.0
pytest-django>=4.7
pytest-cov>=4.1
ruff>=0.8.0
pre-commit>=4.0.0
